+++ NEWS.md
@@ -7,6 +7,7 @@ New features:
 * Added default label color values to all built-in data as a new `lab_color` column
   * The label colors are either white or black, whichever has better contrast with the background color based on [recommendations by the International Telecommunication Union](https://www.itu.int/rec/R-REC-BT.601-7-201103-I/en)
 * `get_scale_data()` now returns label colors (as above) for Macrostrat timescales
+* `facet_wrap_color()` and `facet_grid_color()` now automatically set strip labels to white or black (as above)
 
 Bug fixes: